split tree addon
