+++ common/addons/generation-stage-tree/src/main/java/com/dfsek/terra/addons/generation/tree/TreeLayer.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.addons.tree.tree;
+package com.dfsek.terra.addons.generation.tree;
 
 import com.dfsek.terra.api.block.state.BlockState;
 import com.dfsek.terra.api.noise.NoiseSampler;
