+++ common/addons/generation-stage-tree/src/main/java/com/dfsek/terra/addons/generation/tree/TreeLayerTemplate.java
@@ -1,12 +1,10 @@
-package com.dfsek.terra.addons.tree.tree;
+package com.dfsek.terra.addons.generation.tree;
 
 import com.dfsek.tectonic.annotations.Value;
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
-import com.dfsek.terra.api.structure.Structure;
 import com.dfsek.terra.api.util.Range;
 import com.dfsek.terra.api.util.collection.ProbabilityCollection;
 import com.dfsek.terra.api.util.seeded.NoiseSeeded;
-import com.dfsek.terra.api.world.Flora;
 import com.dfsek.terra.api.world.Tree;
 
 public class TreeLayerTemplate implements ObjectTemplate<TreeLayer> {
